requirements.txt
matplotlib==3.3.4
streamlit==0.76.0
numpy
opencv_python==4.5.1.48
Pillow==8.1.0
